+++ resource/chatadapter.go
@@ -77,14 +77,16 @@ func gChatMessage(
 
 	var icon string
 	switch state {
+	case abortState:
+		icon = "🟤"
+	case errorState:
+		icon = "🟠"
+	case failureState:
+		icon = "🔴"
 	case pendingState:
 		icon = "🟡"
 	case successState:
 		icon = "🟢"
-	case failureState:
-		icon = "🔴"
-	case errorState:
-		icon = "🟠"
 	default:
 		icon = "❓"
 	}