+++ resource/chatadapter_test.go
@@ -15,7 +15,7 @@ func TestChatAdapterIntegration(t *testing.T) {
 	if len(gchatUrl) == 0 {
 		t.Skip("Skipping integration test. See CONTRIBUTING for how to enable.")
 	}
-	gitRef := "cafefade"
+	gitRef := "32e4b4f91b"
 	job := "peel"
 
 	// We send multiple messages to better see the UI in GChat and run the subtests
@@ -25,12 +25,24 @@ func TestChatAdapterIntegration(t *testing.T) {
 		state    string
 	}{
 		{
-			pipeline: "avocado",
-			state:    successState,
+			pipeline: "coconut",
+			state:    abortState,
+		},
+		{
+			pipeline: "coconut",
+			state:    errorState,
 		},
 		{
 			pipeline: "coconut",
 			state:    failureState,
+		},
+		{
+			pipeline: "coconut",
+			state:    pendingState,
+		},
+		{
+			pipeline: "coconut",
+			state:    successState,
 		},
 	}
 
@@ -38,7 +50,7 @@ func TestChatAdapterIntegration(t *testing.T) {
 		tc := tc // Capture range variable, needed for t.Parallel to work correctly.
 		t.Run(tc.pipeline, func(t *testing.T) {
 			t.Parallel()
-			buildURL := "https://ci.example.com/invalid"
+			buildURL := "https://cogito.invalid/teams/TEAM/pipelines/PIPELINE/jobs/JOB/builds/42"
 			ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 			defer cancel()